+++ Keas.Mvc/ClientApp/components/Equipment/AssignEquipment.tsx
@@ -3,7 +3,8 @@ import * as React from 'react';
 import DatePicker from 'react-date-picker';
 import { Button, Modal, ModalBody, ModalFooter } from 'reactstrap';
 import { Context } from '../../Context';
-import { IEquipment, IEquipmentAttribute, IPerson, ISpace } from '../../Types';
+import { IEquipment } from '../../models/Equipment';
+import { IPerson, ISpace } from '../../Types';
 import AssignPerson from '../People/AssignPerson';
 import EquipmentEditValues from './EquipmentEditValues';
 import SearchEquipment from './SearchEquipment';
