+++ Keas.Mvc/ClientApp/components/Equipment/DeleteEquipment.tsx
@@ -1,7 +1,7 @@
 ﻿import * as React from 'react';
 import { Button, Modal, ModalBody, ModalFooter } from 'reactstrap';
 import { Context } from '../../Context';
-import { IEquipment } from '../../Types';
+import { IEquipment } from '../../models/Equipment';
 import EquipmentAssignmentValues from './EquipmentAssignmentValues';
 import EquipmentEditValues from './EquipmentEditValues';
