+++ Keas.Mvc/ClientApp/components/Equipment/EditEquipment.tsx
@@ -1,7 +1,8 @@
 import * as React from 'react';
 import { Button, Modal, ModalBody, ModalFooter } from 'reactstrap';
 import { Context } from '../../Context';
-import { IEquipment, IEquipmentAttribute, ISpace } from '../../Types';
+import { IEquipment, IEquipmentAttribute } from '../../models/Equipment';
+import { ISpace } from '../../Types';
 import EquipmentAssignmentValues from './EquipmentAssignmentValues';
 import EquipmentEditValues from './EquipmentEditValues';
