+++ Keas.Mvc/ClientApp/components/Equipment/EquipmentBigFixSearchId.tsx
@@ -1,6 +1,5 @@
 import * as React from 'react';
 import { toast } from 'react-toastify';
-
 import {
   Button,
   Form,
