+++ Keas.Mvc/ClientApp/components/Equipment/EquipmentContainer.tsx
@@ -3,7 +3,8 @@ import { RouteChildrenProps } from 'react-router';
 import { toast } from 'react-toastify';
 import { Button } from 'reactstrap';
 import { Context } from '../../Context';
-import { IEquipment, IMatchParams, IPerson, ISpace } from '../../Types';
+import { IEquipment } from '../../models/Equipment';
+import { IMatchParams, IPerson, ISpace } from '../../Types';
 import { PermissionsUtil } from '../../util/permissions';
 import Denied from '../Shared/Denied';
 import AssignEquipment from './AssignEquipment';
