+++ Keas.Mvc/ClientApp/components/Equipment/EquipmentListItem.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { Button } from 'reactstrap';
-import { IEquipment } from '../../Types';
+import { IEquipment } from '../../models/Equipment';
 import { DateUtil } from '../../util/dates';
 import ListActionsDropdown, { IAction } from '../ListActionsDropdown';
